code/services/smartapp_masterdata_api/requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
django-db-connection-pool==1.0.7
django-yamlfield==1.0.3
django-jsonfield==1.4.1
django-mysql==3.11.1
django-background-tasks==1.2.5
drf-yasg==1.20.0
PyMySQL==1.0.2
djangorestframework-simplejwt==4.6.0
django-request-logging==0.7.2
python-jose==3.2.0
pandas==1.2.2
numpy==1.20.1
django-rest-framework==0.1.0
djangorestframework-csv==2.1.0
django-markdown==0.8.4
django-filter==2.4.0
django-rest-swagger==2.2.0
django-url-filter==0.3.15
djangorestframework~=3.12.2